Reaaded WWISE original ue5 plugin code

+++ Plugins/Wwise/Source/AkAudio/Private/AkReverbDescriptor.cpp
@@ -139,7 +139,7 @@ void UpdateVolumeAndArea(UBodySetup* bodySetup, const FVector& scale, float& vol
 double FAkReverbDescriptor::TriangleArea(const FVector& v1, const FVector& v2, const FVector& v3)
 {
 #if UE_5_0_OR_LATER
-	float Mag = 0.0f;
+	double Mag = 0.0;
 #else
 	float Mag = 0.0f;
 #endif
